Don't inject incorrect device information
The device information injected by Studio is used by the build system to
optimize the build for the given deployment device.

However, various codepaths use AndroidVersion.DEFAULT (= API level 1)
for deployment target device version if the version could not be
determined, so the AndroidDeviceSpec is updated to ensure that if the
API level could not be determined then Studio does not inject it to the
build.

I haven't managed to reproduce the specific issue, but in general it
seems good to be resillient against this, as the build system takes the
injected values at face value, and injecting
android.injected.build.api=1 will cause the build to fail.

It is safe to not inject target device information, it just might
make the build a bit slower.

Bug: 174988700
Test: Expanded AndroidDeviceSpecUtilTest
Change-Id: Ib159676fe560d62ffe19a67805c3526a575538cc

GitOrigin-RevId: 631ac20967d8c64e8803535fc05143030d1dd769