Fix wrong animation duration in animation preview
When opening the animation inspector, we were getting the incorrect
animation duration until we switched states using the UI. That was happening
because in the new time control infrastructure the first calls to both
setClockTime() and getMaxDuration() take longer than the default timeout of
LayoutlibSceneManager#executeCallbacksAndRequestRender (30ms). It's unclear
why the first call to these methods takes longer than the following ones,
but one theory is that the framework lazily hooks up the state observation
logic, i.e. only after the mutable state (targetState for setClockTime() and
totalDurationNanos for getMaxDuration()) is read from or written to.

We're addressing that by using a more generous timeout the first time we set
and/or read these animation states. It's only possible to use this longer
timeout (5s) because we don't need to trigger the aforementioned methods
from EDT.

Fixes: 179510709
Test: ComposePreviewAnimationManager
Change-Id: I60df72354ad361cf6e82bf8ce3dbc87ccbb51917

GitOrigin-RevId: c47fbf8f28d533d288459e70863264780015bd2d