Add parallel RunningDeviceTarget fields to State
1. Cherry pick this commit
2. Select an available (not running) BootWithSnapshotTarget with the
   dialog
3. Deploy the app
4. Open the dialog
5. The running device should be selected but it's not. That's a bug.

I'm going to fix it by keeping parallel RunningDeviceTarget fields in
State for the drop down and the dialog. If the dialog target is running,
I will return a RunningDeviceTarget instead of the original
BootWithSnapshotTarget. That will fix Step 5 in the workflow above.
Keeping the RunningDeviceTargets separate prevents the original
selections (still in the old fields) from being clobbered.

This is Commit 2 of several commits. The commented out code in
VirtualDevice was me being too eager with Commit 1.

Bug: 179836372
Test: Existing tests
Change-Id: Idbfa3983777a095e8657d457390c309873a1a298