[Dsl] Treat subprojects properties as applied
Properties from a parent's allprojects { ... } or subprojects { ... }
are similar to properties from applied files: they're not actually on
file for the file in question, even though they have an effect on the
Dsl object.  Instead of using setParsedElement() for application of
subprojects properties (and the subprojects block itself, for
propagation to children), use addAppliedProperty() which now needs to be
public.

We also need to replicate the logic in setParsedElement which handles
top-level Java language-level properties in Groovy, inserting a
synthetic JavaDslElement to hold them.  This too should be treated as
APPLIED rather than EXISTING in subprojects to which it is copied.

If we then query the position of Plugins and Apply blocks using
getCurrentElements() rather than getAllElements(), the positions we
compute correspond to the (eventual) positions of the blocks in the Dsl
file, so the consideration of the 0th element becomes more valid.  This
change is also necessary in a few other places, where the distinction
between getCurrentElements() and getAllElements() (which has always been
there but has been less easy to expose with normal Gradle build files)
was not appreciated.

Bug: 147139838
Fixes: 174827009
Test: added to PropertyOrderTest
Change-Id: I202cd3f0acd0ceb96a4df4352c091cb7802f8ab8

GitOrigin-RevId: cc1edac9c6bfe49c3149009f9b91cd2f306390aa