Fix tab border not showing in Animation Preview
We were using a TabbedPane in Animation Preview, displaying one tab per
animation. However, this Component's default behavior is to hide the tab
border (blue line on its bottom) if there is only one tab in the panel. We
now replaced the TabbedPane with JBTabsImpl, which allows to set a custom
tab border, therefore fixing the issue.

Bug: 179684216
Test: ComposePreviewAnimationManagerTest + manually tested visual changes.
Change-Id: I7a7ced743ceb8ca4af478b970218dd87cf67d531