Better out of date lookup for Compose Preview
This CL changes how the preview detects the out of date file in the
editor. Before this CL, the preview would rely on the last modification
timestamp of the VirtualFile to provide the notification. This would
trigger for every change but only once the content has been saved.
It also would trigger for every change.

This CL does a better file analysis and will trigger the notification
immediately for relevant changes.

Test: Added tests
Bug: 176889969
Change-Id: Ifc7820885e9741b81851a8e4bcd50cbe5828f468

GitOrigin-RevId: 5adfc328842b53cd7d09ab0cf6516050dee8fc9f