Fix TargetsForReadingSupplier
State.runningDeviceTargetSelectedWithDropDown and
runningDeviceTargetsSelectedWithDialog are for RunningDeviceTargets. And
RunningDeviceTargets are for virtual devices that don't know if they
were cold booted, quick booted, or booted with a snapshot. Until we get
a new emulator, that is all running virtual devices.

All this fuss continues to work around the scenario in the commit
message for 92ad72d5a21130064145ae9bcdbac3a176ec28b1

Line 51 in the old version of TargetsForReadingSupplier was dropping
RunningDeviceTargets too eagerly. We only want to do it if there's a
nonRunningDeviceTarget in the targets argument (which ultimately comes
from State.targetSelectedWithDropDown and targetsSelectedWithDialog) to
fall back to.

Swapping out the map collected at Line 47 in the old version with the
new isDeviceRunning method serves two purposes:

1. To avoid NullPointerExceptions from calling isConnected on null map
   values
2. To be consistent with the stream approach taken at Line 60 in the new
   version

This is Commit 6 of several commits

Bug: 179836372
Test: TargetsForReadingSupplierTest
Change-Id: Iee24ed8e61e1c5126fd4f413b9bcf9325e2df5db

GitOrigin-RevId: 10e08947810a12f0e12497055053ad48d43687be