Choose Entity for Dao/DatabaseView with respect to Database.
If there are two or more tables with the same name we can resolve to the wrong one.

We're trying our best for resolving to the right one in Dao/DatabaseView:
Search for Databases that return the Dao/contains DatabaseView and look for entities attribute in @Database annotation.

If there is no such Database returns arbitrarily table as before CL.

Additionally fix @Database annotation, replace `tables` attr with `entities`

Fixes: 151239595
Test: test_duplicatedTableNames
Change-Id: I0bcb010d5b72a107d07b2e5afe1e6bbe49c073ad

GitOrigin-RevId: 3731c35070b1a7eb2a08d18c4b4e3dd9ec0290b0