Update ui-tooling auto-import reference to the new namespace
Since Compose alpha08, the ui-tooling module (among others) was moved to
the androidx.compose.ui namespace, as opposed to the previous androidx.ui.
We need to suggest the new namespace in the auto-import feature, otherwise
users will add the obsolete artifact to their build.gradle pointing to
alpha07, which is the latest version of the artifact in the old namespace.

Fixes: 180944692
Test: Manual
Change-Id: I21bbeee8b270c98248949a5a97f21f3dc2484391

GitOrigin-RevId: 758dbfd5f286cf24523c8f74966aec3662689be9