[Dsl] use ModelEffectDescription for new reset()
By ensuring that the new reset element is always created with an
appropriate ModelEffectDescription, we do not need to search for
elements by name, but can instead search for the effect on a given
property.

This also requires a minor tidyup of the implementation of resValues and
buildConfigFields in FlavorTypeModel, again to ensure that the syntax of
the created elements matches closely enough the result of parsing those
elements.  We cannot yet remove all uses of String names for those
properties because the interfaces in GradleDslElement have not yet all
been defined.

Bug: n/a
Test: existing
Change-Id: Ia45cb2a989b15826c311ba6522b4231513903080

GitOrigin-RevId: 5bd8aa47dc5c77ecd30c86120eceaea6c24e284b