Add additional constraint on emulator SD card field
The old property setup was validating only when the external SD card
text field was changed. So if the user sets an invalid SD card location
but then opts to not use the SD card, the validator wouldn't run, thus
leaving around a stale invalidate. This resulted in the user not being
able to finish the config.

Fixes: 128106333
Test: select and  set an invalid external SD card location and change selection
Change-Id: I4ce3b6b9d0c5209b27990283365ece67928c657f