Limit string length in network raw data viewer
The raw data viewer in network profiler uses a JTextArea, which
may freeze the UI thread when its text is too large. As a stop gap,
this change will limit the string length. By default, the pretty
data viewer is displayed so this shouldn't take away users' ability
to view the entire payload.

Fixes: 170799243
Fixes: 131854737
Bug: 180949197
Test: added unit test and manually verified no freeze
Change-Id: Ib01455c792b116f8a4b92a8237768874f49d107a

GitOrigin-RevId: 4c30b13db8ae0e0a6834733f6813bafacd2e68c5