Prepare variant syncing for parallel sync
Parallel sync relies on ability to run multiple build actions in
parallel. However to have multiple build actions to run we need to know
which build variant to request for each module upfront and this is not
possible without first syncing modules depending on it.

In this change we:

(1) restructure variant syncing to process all modules for which we know
    the variant to request at once, for example:

    => :app[debug]
      => :lib1[debug], lib2[debug]
        => :lib1-dep1[debug], :lib1-dep2[debug], :lib2-dep[debug]

(2) when re-syncing an already synced project and build action parallel
    execution is available we speculatively pre-fetch previously
    selected variants upfront and thus make it possible to request them
    in parallel.

    This optimization however would never run with this change since
    the build action parallel execution is not yet available.

Bug: 179156667
Test: existing
Change-Id: Ibd5d2ae303328ec1f8fd40ead6a008bba2dcdc2a

GitOrigin-RevId: 0bf3b6d89973771e904edd2eac675bd09016c796