[Dsl] Allow setValue with list inputs
With atomic properties, the natural sequence to move a property to a
different place is to.setValue(from.getValue()).  It did not work on
lists, possibly because a decision has to be made as to whether to
support a list of property models (as returned from getValue(LIST_TYPE))
or a list of Objects or Strings (as might be natural to do by
construction).

Still, we need this, so be expedient for the Upgrade Assistant and
accept a list of property models.

Bug: 140406102
Bug: 181748057
Test: added unit test; exercised by AUA
Change-Id: If60f0ddd022aa2230d49e78a1fad21660219fa60

GitOrigin-RevId: 0359708e8f1a08c21039efcded68be4852a98820