Special case single translation file with irregular path name
When the translation editor creates a new, non-preexisting, translation
for a particular key, the only piece of information it has is its target
translation language. Therefore, by default it just creates a file at a
specific location (i.e. "values-<lang_code>/strings.xml") if it doesn't
already exist and serializes the newly translated value in that file.

This causes a user expectation mismatch: if the user provided a
translation file that doesn't match the directory format as indicated
above, a new file will be created in the location indicated above and
all newly added translations will be serialized there, instead of the
file the user provided.

This fix works around this issue when there is only one file that's
used as the source translation for a particular locale. It does so by
crawling through all translated keys and checking their translation
source file. When there is only one unique file, the translation editor
will use this file as the destination file to serialize to.

Test: TranslationsEditorSavingTest.java
Fixes: 128106334
Change-Id: Ic00ec378f79512fc901f8eb6dccaf3dc31754519

GitOrigin-RevId: 0ac6066a752c2b94a6f5d8dc8c922a55fcf1e925