[BA] Fix CC warnings number calculation
NoIncompatiblePlugins state should be counted as 1 warning as
it is still a problem that user doesn't have CC turned on.
We add this as a warning to warnings list.
Making it 0 leads to a bug  when we don't show it if
there are no other warnings in the build, because total
number of warnings is 0 and we hide the panel completely.

Bug: 173784116
Test: existing
Change-Id: I51daddf08bab2a19c9f64190315f0bff46363156

GitOrigin-RevId: 2e60fff7482a8c9baa8e09e43d47c16783d30b2b