[AUA] Preserve the order of found usages
The main path through BaseRefactoringProcessor preserves the order of
usages from findUsages() through to execute() / doRefactoring().
However, the path through preview calls appendUsagesInBulk() which adds
usages to a ConcurrentHashMap, so subsequent retrievals of the
non-excluded usages from the preview (whether any have in fact been
excluded or not) will come in an arbitrary order.

This matters because we are aiming to construct generic refactoring
processors for API changes, and those changes might involve additions,
moves and removals.  These operations will have to be performed in
order (otherwise, for example, a block might be removed before its
properties have been moved elsewhere); it is straightforward to reason
about ordering if the order is preserved from the Refactoring Processor
through the entire process, and much less so if we need to reconstruct
order after the fact.

(We can temporarily get away with not removing blocks, because empty
ones will be removed automatically -- but at the point when those Dsl
entities are no longer supported by AGP, any stray text causing a block
to be retained, such as unparsed statements or comments, will cause sync
failures).

Bug: 178092637
Bug: 181768979
Test: manual
Change-Id: I35ae09817461f4febfc55453e95d78ee3972055c

GitOrigin-RevId: 671f50c3145ec62d1bdc771c75d6cd2d3dfc2cc9