Support parameter element references
Add a lookup of parameter elements to ComposeParametersCache.
When an empty parameter group with a reference is expanded in the
UI, we lookup the reference in the cache.

If found we clone the sub elements before expanding the group.
Otherwise the parameter element is loaded from the agent and the
result is added to the sub element before expanding the group.

Bug: 180952854
Test: Added to AppInspectionPropertiesProviderTest
Change-Id: Ia9f9100beb841b39b354128cf5973411e9495389

GitOrigin-RevId: 36719f80e0e27f6f1586fae6fb29ccbdd8aff32a