Make split controls configurable for kotlin files
We recently introduced the split editor to all Kotlin files for compose
projects, attempting to improve the discoverability of the Compose Preview
feature. However, some users are still expecting to open such files in Code
mode by default, even if they contain @Composable functions. For that
reason, we need to allow users to select the default mode for kotlin files
that contain @Composable functions in them and for other kotlin files,
similar to what we do with Design files.

Bug: 181593654
Fixes: 179917375
Test: ComposePreviewRepresentationProviderTest#testDefaultLayout
Change-Id: I1994ec00471e3a648b96b9f45027e12ae23360df

GitOrigin-RevId: b7d7ac0d2e243b683be43a3d3498058e758742f2