LL: Don't crash when there is no namespace.
The editor should never tell LL monitor that a
literal is in function "Fun" without a class
like "ClassName.Fun".

It is possible that the file is in the middle
of editting and therefore does not have valid
syntax so the parser is confused.

We should at least not crash in the case.

Bug: 181779009
Test: AndroidLiveLiteralDeployMonitorTest
Change-Id: I3d8901b62b901f14045940c8bd7a7d801bbce434

GitOrigin-RevId: 076d226917dd16d201027123466bcd8fb29e853c