LL: Don't crash when AndroidSessionInfo is null
It might appear that some ProcessHandler does not have a
corresponding AndroidSessionInfo in the project.

I do not know the exact cause of this, may be the project
has some non Android part that can it can run like host
unit tests or a main().

This CL will filter out nulls from the list of all
AndroidSessionInfo so caller can assume all entries are
non-null.

Bug: 181684039
Test: Manually verified LL still works.
Change-Id: If8e5b649eee24c549f46794822f0d3b4200cfdb5

GitOrigin-RevId: 1d65b6f8c34572a6ff5a15b9c74490408bfa97e5