[MPSS] Ensure Android data services can correctly set SDKs
This change ensures that Android sdk setup runs after that of the Kotlin
plugin. This is since the Kotlin plugin uses the presence of an
AndroidFacet on the module to determine the need to setup the sdk. With
MPSS this facet is only on the main module so other modules sdks get
overriden.

Bug: N/A
Test: N/A
Change-Id: I54b77dcf245ea30890e38dfda14f0615f922245e

GitOrigin-RevId: b4a957cd7ee0abe8ee79b9b2931d3ee2d40e5745