Check in network inspector front end code
This change forks network profiler's frontend UI and model code into the
network inspector module and converts them to Kotlin.

Certain interfaces needed to be modified and new ones were created, but
most of the code (especially the UI components) are unaltered.

With this change, we will have a fully functioning network inspector
checked in.

Test: To be written in future change. Some can be migrated.
Bug: 178232811
Change-Id: I07f06ca502c6170e80f10759e2c90321df77c19b

GitOrigin-RevId: f49b47041a0f52e2ca5c2d726aa0125330a52a77