Change how we know if we can rotate
Previously we allowed rotation if the rendering had more than one image
in it. This was just a heuristic, since we could have an app that only
rendered in one rendernode.
Now, if we're going to render a flat bitmap when not rotated, we need to
keep track of whether it's possible to show a rotated view separately,
so using the existing capabilities mechanism for this.
Note that this makes the capabilities list mutable where previously it
was not.

Test: in practice there's not really any behavior change, but maybe
      there should be a new test anyway. I will look into it later,
      since there's some time pressure for these changes.
Bug: none
Change-Id: I218921134d04ef645c347f1e21980cc4ec0ce7e8

GitOrigin-RevId: 53507246c62853aef2d6b7da1000c0d4f490b465