Reset animation preview clock time after build
When building or trigerring a refresh after changing a live literal, we
update the animation preview tabs so they represent the updated animations.
However, we were not reseting the clock, so both the clock time and animated
properties did not correspond to the animation being previewed until the user
interacted with the animation preview (e.g. clicking on the timeline or
changing the from/to states). This change fixes the issue by resetting the
clock when resubscribing the animations.

Bug: 179511256
Test: Manual by refreshing the preview with the animation panel open
Change-Id: I1b57684edcd25f8df4ba9a0b6e392b6723f6c3c2