[AUA] Add error processing to version selection field
Highlight version field with an error when it contains empty text,
text that can not be parsed to a valid AGP version or parsed version is
lower than the current version.
Pass only valid versions to the tool window model otherwise pass null.
In case of null version in the model show empty state (just for now,
add better error state messages later).

Bug: 178569506
Test: N/A
Change-Id: I518db038d70fc648277f2ccdba8974a8136a8f10

GitOrigin-RevId: fe6098bacb2d721a17be7cdc42675128c4cb9e00