[Kotlin] Handle all language features through Dsl model
Although there are apparently no other current callers of general
feature configuration, there is the possibility that third-party plugins
might use this to configure their own language feature requirements (or
to provide UI for the user to modify the language features themselves),
and any attempt to do so in an Android Kotlin module will fail with the
default build file modifications.  Use our Dsl model of kotlinOptions
for all language features and for both ENABLED and DISABLED state, and
make the lack of support for ENABLED_WITH_WARNING and ENABLED_WITH_ERROR
explicit.

Bug: n/a
Test: added
Change-Id: I35d129b87eb92be752cfe6101f90d8a82f37ce7c

GitOrigin-RevId: 455a76f6828411a9576d37211e9f6e42c6a63104