Add a weak warning if updateTransition label is not set
We use the `label` parameter when creating Compose transition animations to
help users to identify the transition in the animation preview. Since this
parameter is null by default, we need to show a weak warning so users are
aware that they should be using that to make the most out of animation
tooling.

In this change, we also remove the obsolete PropKeys weak warning, since
PropKeys based transitions (v1) are now deprecated and removed.

Bug: 180492051
Test: AnimationInspectionsTest
Change-Id: I563bf1bad5d7e4f77d9973cc543c33df7e4db8d9

GitOrigin-RevId: 9d0c441469b9e56bb05ec5066eec965be76d225d