Remove false assertion when updating delta instances
This change removes the wrong expectation similar to one for updating
snapshot instances, complicated by the fact that the function is
already returning a boolean indicating whether the instance was been
added/removed from the set. We extend the functionn to return a
boolean pair (conceptually), but use an enum to document that only 3
of the 4 combinations make sense.

Fixes: 179172432
Test: n/a
Change-Id: Ie38c1675e23ce4789bf1e053daaee33bd955a968

GitOrigin-RevId: 2643dc9c30edb3e882fe9fa8a386699c62f282f1