Fix IndexOutOfBounds exception in thread state table
Move the thread state distribution logic to the main thread. This
should avoid having stale table model that causes IndexOutOfBounds.

Tracing the code shows that on a 200 MB trace with 3000 thread
states, it takes less than 1 ms (200 us on average) to do the
computation. So moving it to the main thread shouldn't freeze the
UI.

Fixes: 175675564
Test: manually to verify no more exceptions
Change-Id: I2117e6ce681c84a9b279832b8a7f80d2d207c6b6

GitOrigin-RevId: bd7af435d75a6d0199356b61209f52268b388f67