Provide auto-completion when managing dependencies
... in 1) build.gradle/build.gradle.kts files
2) java/kotlin code in root project directory
"buildSrc".

For 2), users might introduce/edit java/kotlin code
before initial gradle sync, i.e gradle facet
might not be there yet, so we just check the
pattern of the file path for condition check.

Same for 1), that for simplicity, we don't check if
it's a gradle module or not, instead, we only check
file extensions.

Bug: 134753077
Test: Added
Change-Id: I1857517fc784ed2a0dc99495d9066416e26bccf2

GitOrigin-RevId: 0c299735173efe52560209646759af7391c2513f