Add border to Compose Preview toolbar DropDownActions
Blueprint mode action and switch layout action are both DropDownActions, and
as such do not have the default empty border, which gives a 2 horizontal
padding and 1px vertical padding. ForceCompileAndRefreshAction is AnAction
and for that reason does not need the empty border to be defined, since this
is already done in ActionToolbarImpl#createToolbarButton.

Fixes: 182587148
Test: Manual/Visual.
Change-Id: Idd7bc672a908bd91f15d6cc40c64941f7999b66c