Set Android Support version we use in tests to "dev build"
This is to stop tests from using the platform version to sent to AGP
which can cause issues as AGP validates this. The current version set in
tests should never be present in production and may cause some tests to
fail. We change this to "dev build" which prevents the sending of the
AGP version in tests.

Bug: N/A
Test: N/A
Change-Id: I6f39da6fa26df662bb90242f5d23d251e084972d

GitOrigin-RevId: 1e651fb4d9ccc438d17e871e67c3e891c5f1ef45