Force scrolling when navigating to @Preview
When clicking on a @Preview annotation, we navigate to the Compose Preview
in the design surface. However, if the preview is only partially visible, we
don't scroll at all. This CL makes sure we scroll the surface to a position
that shows the top of the Compose Preview corresponding to the @Preview
annotation clicked by the user.

Fixes: 182571820
Test: Manual. Click on a @Preview annotation and verify Preview is visible.
Change-Id: Ie2a5bfa97bc7159e35db7e5be68cbb34eab09283