Repaint the animation preview when there are no animations
The animation panel might not have the focus when the "No animations" panel
is displayed, i.e. when a live literal is changed in the editor and we need
to refresh the animation preview so it displays the most up-to-date
animations. For that reason, we need to make sure the animation panel is
repainted correctly.

Bug: 179511256
Test: Manual
Change-Id: I752c89b3dfce2e85e4397bb01d78bb7f9280172d

GitOrigin-RevId: 04fae7b968da13bf51e23a46000b495933840140