[AUA] Test the initial state of the treeModel
In order to test the state meaningfully, we need some active refactoring
processors, which means we need some non-empty files in the project.
Change the way we set up these tests to make that a bit more
straightforward, using AndroidProjectRule rather than
LightFixtureTestCase.  Then test that if we have a build.gradle file in
the root of the project with an upgradable AGP classpath declaration, we
see an active MANDATORY_CODEPENDENT processor with the right title in
the treeModel.

Bug: 183107211
Test: added here
Change-Id: Ib35a25d2046eb69ef1a3d46ac22e1c62661cd51d

GitOrigin-RevId: cae455d01d7eac0db7a8732df6ada05835c46eea