Don't request focus if split editor state is redundantly set
SplitEditorAction#setSelected is called by the IntelliJ actions mechanism
even if the state being set is equal to the current state. Since we request
the focus to the editor when *changing* modes in this method, we were
incorrectly requesting the focus even when there was no mode change. This CL
fixes this issue by returning early when SplitEditorAction#setSelected is
called with the currently selected mode.

Fixes: 182682222
Fixes: 183061941
Test: Manual (test scenario in the bug) + DesignToolsSplitEditorTest
Change-Id: Ifa44f44be7ed89d0ee45a09f0f870f2b2cb4b29c