Move early return after delegate Action is called
In SplitEditorAction#setSelected, we return early when redundantly setting
the state. However, we should still call the delegate's setSelected to let
the IntelliJ action handle the delegate's logic on its own. If we don't do
that, it can happen that a file is open in split mode but the preview panel
is not displayed until we change modes back and forth (e.g. change to code
and back to split).

Bug: 183061941
Test: Manual. Open a file in split mode and check the preview panel shows up
Change-Id: Ib0cfda091362977b827ac121d3d8d574189e0ff6