Show animation preview tabbed pane after tabs are populated
It takes some time (one frame in Compose) between subscribing a transition
animation and being able to obtain the correct animation duration and
transition property values. Since we're adding the tab corresponding to the
animation as soon as we subscribed it, the visual feedback for users of
animation preview would be displaying the tab without all the information
and after a split second we would change the timeline max value and
populate the transition properties panel appropriately.

This "blink" in the UI is not great, as users might not know when it's about
to happen. In this CL, we delay the tab addition until the animation
duration and transition property values are available, so the users will see
the "Loading Animations..." panel until the animations are actually ready to
be interacted with, displaying the correct values in the panel.

Bug: 179511256
Test: ComposePreviewAnimationManagerTest
Change-Id: I92117832d87d607a3706070f95b3e516b352fc0b

GitOrigin-RevId: 0ac8cd5b9a2356b64b697879f210d628555a0f37