[AUA] Disable upgrade if the classpath dependency isn't found
As with the dialog box version of this functionality, if we can't find
the AGP version to upgrade it, upgrading the rest of it is unlikely to
work, so disable the upgrade (except if the current and new versions are
the same, in which case it is safe to perform optional steps.  This
happens to be the first time we initialize isNoOpForProject off the
writeThread, so make sure we have read access for that.

In theory we could allow the user to do pre-upgrade steps for their
chosen non-trivial upgrade.  This would involve unchecking and disabling
the MANDATORY_CODEPENDENT tickybox, and providing some indication
somewhere in the UI for the reason that that tickybox is disabled.

Bug: 167615343
Test: added
Change-Id: Iff4d12f8da89d86f5cfa386c39462fbb2fa5e5d9

GitOrigin-RevId: 481800fea6eaf449fbeb5d443809228c5c67836b