[AUA] Update strings referring to version on refresh
The refresh might be triggered by the user wishing to reinitialize the
upgrade process after making some manual changes, including possibly a
manual change to the version of the project.  This change does not
actually allow the tool window's idea of what the current version is to
be updated, but does perform the UI manoeuvers to re-show the relevant
strings.

Bug: none
Test: manual
Change-Id: I16877235e17b42a164781bd824f0d1bf8e76e306