[AUA] Allow current to be updated in ToolWindowModel
Since we might have the state of the project change while the tool
window is open, including the AGP version of the project, we need to be
able to update the ToolWindowModel's view of the current state.

In principle, the plugin info may be unavailable.  That means that this
field should be nullable, because we do not want to falsely claim that
we know the current version when in fact we don't, but have failed to
update a previously-current version.  Handle the new nullability in
various places, and in the process add a check for the validity of
current relative to the target AGP version.

Finding the new current version of the plugin is potentially a long
operation and is also hard to handle in unit tests.  Only do this when
it is required (currently, when the user presses the Refresh button, and
in future when sync completes).

Bug: none
Test: existing, manual, added strings test
Change-Id: Id5d2e2a2606977914d852b082fd166a821562883