Adjust painting of filtered nodes
After this change:
- borders of filtered views are omitted
- levels of filtered views are collapsed

The effect is that nodes selected in the component tree that are
drawn by hidden system nodes looks like they are selectable in
the image.
Example: A Text node can be selected in the inspector by clicking
on its text, although the text is actually is drawn by a Layout
node in a child node of the Text node.

Fixes: 182838650
Test: Added unit tests
Change-Id: Iaa73235e01ba8e7109ae00e4125ea64f8c395bb6

GitOrigin-RevId: f4f165884a7dc8c6e47f95281fd78dff1fe93878