[Sync] Rename and document a mostly-internal function
It is useful for the Upgrade Assistant to know which versions of AGP
should be considered sufficiently-incompatible with the running version
of Studio, so that it can avoid suggesting upgrades to those versions;
it would be poor experience for the user to perform an upgrade, only to
be immediately prompted to upgrade again (and, because of the way force
is implemented, without any affordances beyond a modal dialog).  There's
no particular reason to have this only for tests, except that there's an
implicit assumption that in production `recommended` is always the
version returned by LatestKnownPluginVersionProvider, so document that.

Bug: none
Test: existing
Change-Id: I07d4e158361904c399b94e98e23780352fb5091c

GitOrigin-RevId: ce584da0b9dde29de4eb662af2b4ca96db757e4f