[BA] Remove time percentage from CC warning pages
It does not provide much value to the user but can
imply a very incorrect picture in some corner cases.
For example in case of no-op build it'd show 100%
of total build time even though it can never be that high
because CC will anyway still waste some time on serialization
logic.

Bug: 183590986
Test: N/A
Change-Id: I23387bc2e1c8441d586cb9e5fce82944f018b6b7

GitOrigin-RevId: 4bf75204126668cfcb7d371397eda8d642552ab0