Use the trace's own type instead of the recording config type
There has been a bug where we use the current recording type in the
configuration UI to parse recorded traces. This was fine because we
used to try all trace types during parsing. But a recent change
made the CPU parser pick the known trace type, so now this mismatch
causes parsing failures.

The fix is the right thing to do because we already know the trace
type and should always use that instead of what's selected in the
UI.

Fixes: 138677869
Fixes: 150232730
Test: added
Change-Id: I4374f201eb9ac1059c9052dec676958de76e3f42

GitOrigin-RevId: 441e2213aa45dcb3b399899fab00298183769d27