[AUA] Escape the colon when writing distributionUrl
Apparently not having it causes problems in some circumstances.

We retrieve the current URL from the properties file using a routine
which handles any escapes that might be present already; the URL we
consider is always effectively unescaped.  The updated URL follows the
same form (i.e. has no escapes in memory) so it is always correct to
escape the colon without needing to consider the possibility that it has
already been escaped.

One might hope that the property file writing routine would handle this,
but apparently the writing routine does not escape embedded delimiter
characters.  There apparently exist readers for this property file
format which consider a property line with first an equals sign then an
unescaped colon as something other than the property whose key precedes
the equals and whose value is after.  Existing practice seems to be to
unconditionally escape the colon in distributionUrl lines.

Bug: 183644188
Test: added, existing tests updated
Change-Id: I6f3fb81b580cfd365ea0c0d52db2bf44e878a9cf

GitOrigin-RevId: 256a7f7930bbaf8b7caa90abef6c129df640230f