[AUA] Clarify and fix the filter of known versions
Ensure that the latest known version is present in the list of
suggestions, whether it is known to the Google Maven repository or
not.  (This matters particularly for testing, but is probably also a
good idea in case for whatever reason a user's network cannot reach the
Google Maven servers at a particular time).

Include also both the current version of AGP and the latest known
version of AGP in the known version list.  We include the current
version because "upgrading" to the current version is one way of viewing
optional and pre-upgrade steps for a subsequent version without worrying
about changing the running version of AGP itself.

Bug: 183783618
Test: manual, unit tests in next CL
Change-Id: Ic80b5b39658f819e11847cea0b0c6c9a4209d99d

GitOrigin-RevId: 08d4076d1d678499e68e243501777b588ebfc300