Do not override project-jdk with  IdeSdks.getJdk
... when importing a project, creatting GradleExecutionSettings or
setting default JDK.

For GradleProjectImporter this was needed since we wanted to always use
the same jdk, but now it is possible to have a different JDK per
project, so we should not be changing this value when importing a
project (only if it is a new project).

GradleUtil was replacing the settings generated by
AndroidStudioGradleInstallationManager, but this is no longer necessary
since now this class handles both cases (when using JDK per project and
not).

IdeSdks no longer replces project-jdk of open projects when setting JDK,
but removes invalid JDKs from project table so they are marked as such
when displayed in the JDK combo box.

AndroidGradleTests was modified accordingly to have the same behaviour
(test projects will use embedded JDK as project-jdk).

Bug: 179428268
Test: testJdkPathFromProjectJava8, testJdkPathFromProjectJavaCurrent
Change-Id: I5ab38acd29d26ba9018b80b88303ed243c8d5db9