[Sync] Make GradleFiles use more local read actions
When computing the hashes for various Gradle build files in a project,
we only need a read lock when inspecting the content of a file.  In
particular, we should not hold a read lock when we might refresh the
filesystem, as that can lead to deadlock.  When we used to run this on
the EDT, the deadlock was not possible, but running it on the EDT could
lead to UI freezes (particularly if the filesystem was slow).  Now that
we run this on a background thread, we are potentially exposed to the
threat of deadlock.  Moving the read actions to cover purely the
functionality that needs it removes the threat.

Bug: 177796351
Test: manual, existing
Change-Id: Iaa108a27277258a2f4fc480fce04dd35bcfb7b3b

GitOrigin-RevId: a08e43f43e1925c70e1603cb334cb125b043d431