Add Compose preview refresh time to Event Log
In an attempt to improve Compose Preview refresh perceived speed, we're
adding the time taken to refresh the previews to the Event Log tool window
in Android Studio. For now, we log the following:

* Time taken to find PreviewElements (including pinned elements)
* Total elapsed time taken to refresh all the previews

We might consider a finer granularity or reviewing/removing the current logs
at a later point.

Bug: 183935976
Test: Manual by checking the Event Log tool window.
Change-Id: If42a92aa06a163d4fb46bd822c06b26d5d403c9a