[PSD] Restrict AGP version variable to Project buildscript scope
Most properties, when extracted as variables, can live in the default
scope (Project-level), or the module scope, or the project buildscript.
The AGP version, however, must be available in the project buildscript,
because that is where it is used.

Allow property definitions to specify the required scope for any
variable; use this specification, if present, to override the default
module/project scope when extracting a variable in the Project Structure
Dialog.

Also specify a preferred variable name for the AGP version property.

Fixes: 169127416
Test: manual
Change-Id: I1b39f5688670cfc4aafcf21b0788d78d3d55e2c7

GitOrigin-RevId: 6c053f7c569e230244124180e3d19716faaf0dc1