Use ProgressIndicator APIs when refreshing Compose previews
Use IntelliJ's ProgressIndicator APIs when refreshing Compose previews, so
users are aware that a long task is running.

The goal of that change is to improve the perceived speed to refresh the
previews, since currently we display multiple spinners spread around the
design surface, e.g. in the preview toolbar (individual spinners) and in the
surface toolbar. Developers are used to watch out for the bottom status bar
when long tasks are in progress (e.g. gradle build, starting AVD, etc.), so
moving/centralizing the loading status awareness to that component makes
sense.

Bug: 183935976
Test: Manual
Change-Id: I9cdb641deef2060924e60ada404325aafd8e96c3

GitOrigin-RevId: 6cf5cf31ace694ed778df3664528f4799687b773