[AUA] Update configurations for more Androidish plugins
The `com.android.base` plugin is a no-op plugin but indicates that the
module has Android nature, so treat modules using that plugin as a
library.

The `android` plugin is a deprecated equivalent to the
`com.android.application` plugin, so treat modules using that plugin as
an application.

Include a test checking that if a module has both a library-ish plugin
and an application-ish plugin (e.g. the two above), it is treated as an
application module rather than a library module.

Bug: 149565531
Test: added
Change-Id: I7f37f9c7f8bde25f0ded594acc4db3506d6a2d8e

GitOrigin-RevId: 634cf7e1141deb1f38c015cbfeb0ea10e85cd3a5