Lint: use fresh merged manifest
Using a stale merged manifest leads to bugs. For example, if a user
types some XML code that should trigger a Lint warning, the warning
will not show up for a long time.

Computing a fresh merged manifest is what Lint used to do in the past.
Still, there are some risks for performance, because computing the
merged manifest is known to be expensive for large projects in certain
cases. On the other hand, Lint runs on a background thread under a
cancelable read action. We can monitor analytics to see if Lint
performance suffers too much after this change.

For reference, manifest merging is relatively fast for the sample
project I've used. E.g. for Tachiyomi it takes around 100 ms.

Bug: 181341721
Test: manual
Change-Id: I9e50e91569348fe17172f9ce84c0b8776f2549c9