Update layout inspector dialog message to include stop action
Previously, if you minimized an active layout inspector,  we would
show a popup dialog informing the user that if they wanted to stop
layout inspection, they should restore the window and select "stop"
from the process pulldown.

Now, we provide a hyperlink to stop inspection directly without
the extra step.

Bug: 184046383
Test: Updated
Change-Id: Ic28dea86c29a1364266cd59d1e4baa61b9a152cc

GitOrigin-RevId: 742f61895c2f0e60ca138ccd654ce1a9ae180da9