Remove spinners and disable preview toolbar when refreshing
We were displaying a spinner in each Compose Preview toolbar to indicate
which previews were refreshing. It seems those individual spinners caused
more confusion and degrade the refresh perceived speed. Instead, we're now
removing the individual spinners and disabling all the toolbar actions.

Now, users will have two pieces of visual feedback to indicate a refresh is
in progress: the single spinner on the top-right corner that indicates the
preview status; and the progress indicator located in Android Studio's
global status bar, located on the center-bottom of the IDE. In addition to
that, they'll realize they can't interact with the preview actions while
refresh is in progress since the toolbar is now disabled, both the global
preview toolbar and the individual preview toolbar.

Bug: 183935976
Test: Manual by refreshing/building the preview and observing the toolbars
Change-Id: I149a39799ce0b0ecd9c2dfe80955abb6bd65d2df

GitOrigin-RevId: c026a31e5549d502242fb26239cadf325c71404a