Store default JDK in ProjectGradleJdk
When a new project is created we use the result of IdeSdks.getJdk() as
the initial value for the Gradle JDK. This default is set when the first
run wizard is executed.

Before being able to use a JDK per project, the wizard would call
IdeSdks.setJdkPath to configure this value. This function would remove all
jdks in the ProjectJdkTable and then add it, effectively causing a
single JDK in the table. When IdeSdk.getJdk was called, it looks for the
first compatible JDK, but since there was a single one, the result was
the expected one.

Now with JDK per project we no longer clean the ProjectJdkTable before
adding a new one, causing IdeSdks.getJdk to return the previous one.

This change fixes this issue by adding "Android Studio default JDK" to
the jdk table and causing IdeSdks.getSdk() to look for it.

Bug: 184667429
Test: Existing tests that use IdeSdks, new test for default JDK
Change-Id: Id75c500fb35c9ecd8306ed36348985d9df11b509

GitOrigin-RevId: 24dfd22caebb031ca46789a044afa01386ba909e