Allow tagging `CaptureDurationData` with more precise capture type
`CaptureDurationData` has been having an ad-hoc property `isSeparateStageData`,
used by the memory profiler to decide to open it in a separate stage.

This change generalizes the ad-hoc field to allow optionally tagging
`CaptureDurationData` with a more precise upperbound of the capture
object, and it's up to the duration data's handler to decide what
to do with the more precise information.

Compared with other obvious approaches (e.g. using enum or constant), using
the class as a tag has the advantage of guarantee from the typechecker
that it's consistent with the capture type returned by
`CaptureDurationData.getCaptureEntry().getCaptureObject()` (one can provide
an imprecise bound, but not an inconsistent one).

Bug: n/a
Test: existing
Change-Id: Ifc4b1c265b0d508b26c1097656baa3ccfe550a74