[Preview picker] Fix fontScale editing
Previously, editing fontScale would assume that the user would input the
value as a float the same way is done in code. And so if they just typed
an scale without the float suffix, unexpected behavior would follow.

This makes it so it's possible to type float numbers in several ways and
the actual value in the file will be in the correct format.

Bug: n/a
Test: PsiPickerTests
Change-Id: Ia70430f9d8b21dbcd7d1c205b199471f7bebc18e