Use longest valid path instead of first valid path
... when paths have spaces while adding links in console output.

We currently stop parsing when the current path contains an space and is
valid, for example, if the path in the error is

"e: /Andriod Studio Project/path/to/file.java: (10:1)"

But the path "/Android Studio" is also valid, the hyperlink only covers
"/Android Studio" instead of the whole file path.

This change makes the parser break a path with spaces only if the path
until the first separator (either '/' or '\') is found and it is not a
valid path.

Bug: 167701951
Test: Added test in GenericFileFilterTest
Change-Id: I18a3f32fce75f8b1911cdda07412149881654d18

GitOrigin-RevId: 8b631ebafdb7f0b7d9710899145eff71f37e659f