Fix live literals action not showing up after hiding it
After hiding the live literals action from the bottom-right status bar in
Android Studio (right-click the bar and uncheck "Live Literals Tracking
Widget"), we were not bringing the action back even if we tried to check the
option back, unless the user triggered a build. We should allow users to
hide and show the live literals action icon without requiring a new build.

Fixes: 182611318
Test: Manual
Change-Id: I0e5af2a7921977e5bcecd5aa1be3a0686446600d

GitOrigin-RevId: 83c1384a9fa7fe2bbcfcf930f4124151a6259655