Don't ignore a new scan request if scanning has already started
It is safe to ignore a scan request only if there is a previous scan request for the same file and the scanning hasn't started yet. In order to implement this behavior we keep track of the running scans separately from the waiting scan requests.

An incremental update of the contents of a resource file is safe only when there is no concurrently running scan.

The CL also includes some cleanup.

Test: existing
Bug: 167583128
Change-Id: I4cb80956ad11a94824819471720d30153e6a5ef5

GitOrigin-RevId: 7293ba922ddc305a378f9e5efaa581470f335c14