[suggested import] Support undo and redo action
It's fixed by registering the undoable action(gradle sync).

After ag/3451476, changes are picked up only when the
corresponding build file is selected in the editor.
Thus, our changes from intention action can't be picked up
natually, and an explict gradle sync is needed.

When the intention action is executed, detailed sub-actions
like editor change actions are collected in the list. And
this is why we can perform undo and redo when requested.
In this Cl, we actually register gradle sync action both
at the begining and at the last to guarantee sync is
requested after all the changes in the build files, regardless
of redo or undo.

Also, when requesting sync, we also explicitly handle those
"sync in progress" cases by guaranteeing another sync after
the current one is done, to ensure freshness.

Bug: 184770348
Test: Updated
Change-Id: I28941f76161cc2b1879ce560efbdea03ed1d8bb4