Make AndroidLogFilterModel a Concrete Class
There's really no need for it to be abstract. The only impls are the
production impl and the test impl. The abstraction is used to hide the
use of AndroidLogcatPreferences in the test impl but that doesn't seem
to be necessary.

This change will make a downstream change a bit nicer since we will be
able to construct a new instance of AndroidLogFilterModel directly with
a single line rather that creating an inlined impl of the abstract
class.

Test: Existing tests provide coverage.
Bug: 157171025

Change-Id: Ic7aea7f528356aaa66a7ad435385d161eaa280fc

GitOrigin-RevId: 2a018a9e66e4325625668f3c9fd3d94123f2eaf0