Disable kotlin compiler version check
This disables the check only for the test to ensure that we can use the
current version of the Compose libraries with the 1.4.31 compiler.
This will be removed after we update all the Compose libraries.

We also disable the template test until we have a way to disable this
check for that test.

Bug: 182477583
Test: Tests pass
Change-Id: Ie7c5bccb45bebef6d0f3ef7e2ade0ac09f4130c0

GitOrigin-RevId: 841ca2989abcc6eb4cf4bade83b7c960efdb981b