[tools/adt/idea] Remove The Deprecated LogCatTimestamp Field of LogCatHeader
The origin of the duplicity is that there are 2 ways of parsing a timestamp.

One way using a "MM-DD HH:MM:SS.mmm" format and another using an Epoch integer. When parsing from an Epoch, we store an Instant but when parsing from date-time, we store a LogCatTimestamp instead.

We can get rid of the LogCatTimestamp in the header if we simply parse the date-time into an Instant. The date-time string is missing a year so we have to "make one up".

Using a single representation of the timestamp also means we don't have to have 2 ways to format it into a string. We current have 1 way to format from an Instant and another to format from a LogCatTimestamp.

Bug: 185479433
Test: Existing tests provide coverage
Change-Id: I752abb891e18ed3d617a92ff8621b0aaa5e7d03a

GitOrigin-RevId: fb0dfe7d56076bb8661c8b23aa7dc67f9b1b647d