Show a notification when Gradle JDK is invalid
... and a project is open.

We are currently showing an error and quickfixes but only after sync,
but there could be other issues before syncing, for example unresolved
symbols from java classes.

This change verifies that the project Gradle JDK is valid when opening
it, showing an error if not. It will also display a quickfix to use
the embedded JDK (if possible) and a link to select a JDK. Also methods
that show notifications when project starts where moved to
ProjectNotificationsUtils.

Bug: 185510260
Test: ProjectNotificationsUtilsTest
Change-Id: I9aada1384e7169a358ef43bd7f7abfc00e04181a