[AUA] Display status on beyond-known AGP versions
It's not in principle wrong to attempt to use Studio to a version of AGP
released later, though support for it is not strong; it is, however,
improving.  In particular the case of differing patchlevel releases of
Studio and AGP should be supported (e.g. if Studio knows about version
7.0.0, we must not disallow upgrading AGP to 7.0.2, though showing a
warning might be fair enough.)  On the other hand, upgrading to a major
version of AGP beyond the currently-known one is almost certain to fail;
this could be reviewed in whatever the `C` release of Studio is,
corresponding to AGP 7.2 (currently planned to be the last minor
revision of AGP before 8.0).

Fixes: 184932699
Test: manual, added to existing test
Change-Id: I9b381bbb5abc486e55093958e7a7fddb264f271d