[AUA] Display indication of running state
Give the user feedback that their request to run the processor has been
accepted by making the same UI changes as to a changed AGP: disabling
the buttons and setting the loading state, though here with a custom
string rather than the default.  This means tracking in the model what
the loading string needs to be at any given time; in each case that
string can be set just before the call which results in the loading
state being displayed.

Bug: 185120639
Test: manual
Change-Id: I2d861a6fe5d98a8078356d9befa637e4a774776e