[AUA] Add editing support to version field
Also simplify logic around validation and value commit a bit.
It is still not ideal, for example it value edited manually does not
get propagated to the combo box model so when it opens a popup
it still believes the old value were selected.
It is not clear though at what point this value should be propagated
so leave it for a follow up cleanup when we untangle the data flow
in this area.

Bug: 185179665
Test: N/A
Change-Id: I854eb81cbaf345f8c46aca33d840df4ef48a9ae8

GitOrigin-RevId: aad3891ebbd6d6e5503b5200418f7a44c9c5870d