Fixes scrolling issues with the assistant
The reason the scroll bars move after the page has been rendered,
is that we keep adding text to it. When we add text the JEditorPanes
try to keep the cursor in the same place. By default the cursor is
at the end, and by keeping it at the end we keep scrolling.

Other attempts at fixing this issue have been around restoring scroll bar
positions after they have changed. Some time it was possible to do that
by doing invokeLater, but in the cases where the text was delayed by
seconds that didn't work. Other attempts to make the text smaller
resulted in text not being visible at all.

Instead this change sets the JEditorPanes to not update their cursors
preventing them from auto-scrolling under any changes.

Bug: 152075176
Test: I think this needs to be fully rewritten to be testable.
Change-Id: Ia08259d3afff4fdbfcd7da173212d81d403ed0b6

GitOrigin-RevId: 305dac9a00522c43bacf2148e11f1be242f9b794