Invalidate configuration when composable name is invalid
When we create a Compose Preview run configuration from a @Preview function,
we set the composable FQN as a configuration property. Users can, however,
rename the function and that would cause the configuration to be invalid,
since it'd be pointing to a non-existant method.

We were ignoring this error, and the user could try to run the invalid
configuration, but the app was crashing. Now, we're properly invalidating
the configuration if the composable FQN it points to is not a valid function
annotated with @Preview.

Fixes: 184901679
Test: Manual
Change-Id: I14685ce6756e8257bd13ba0d888a284ed6082793

GitOrigin-RevId: 0f30d1534d4821a5c7d4555901332e70bc13f975