Recreate JDK from table when importing settings
When updating Android Studio from one version to another, we also update
the embedded JDK. This means that if we import the setting from an older
version, the information on the JDK table for the embedded JDK becomes
stale. This was not an issue on previous version since we have been
providing a version of Java 8, but starting with Android Studio 4.2, we
now provide Java 11, which has different classPath entries on the JDK
table, causing unresolved symbols and other problems.

This change updates all the JDKs in the JDK table after importing
settings so the information in it is correct.

Bug: 185562147
Test: Manually checked (will generate a manual test).
Change-Id: I14860ac898983300191285f9f86a3e1fbe9a7f83

GitOrigin-RevId: e6c2bf6087eb9f7a1501ac4a6e1b422ff664ae1c