Pause timeline if there are no animations being inspected
When we refresh the preview either from a user explicit request or through
a literal change, the animation preview will unsubscribe all the current
animations and wait for the updated animations to be subscribed. If in the
meantime the timeline was playing with the "loop" action triggered, we were
continuing to make "set clock" requests to be executed in the layoutlib
render thread every 30ms using a ticker. That would cause a non-trivial
overhead to the render thread, since it would be busy trying to inflate the
new ComposeViewAdapter, subscribing the new animations, etc.

This CL removes this overhead by pausing the timeline, and consequently the
ticker, after unsubscribing all animations.

Bug: 185171839
Test: Manual by refreshing the preview while playing in loop
Change-Id: I4fc1246f8e5dd55284a511e45e0cef9b6a933f88