[BA] Make analyzers know when running CC trial builds
Before this change Build Analyzers backend could not know
if it was a usual build or Configuration cache trial build
triggered from BA CC pages.
To make it possible to distinguish we need to pass a context boolean
variable from the runner back to the analyzers backend.
Variable is set to true when starting the trial build and flipped
back to false on the post build task, same time when showing the dialog.
Since analyzers are running as part of the build in the very end of it
they get the flag set to true if it is a trial build.

The value will be used in the following CL.

Bug: 183707585
Test: manual
Change-Id: I78fed26cfc9dc60dbc0c6a4575462562b20d394e

GitOrigin-RevId: 2c796efc702f3ec314e43ecd0c47d2859f3b4327