Check layout events for reference equality
If we receive many layout events with the same data we end up processing
all of them in order, as long as more events come in while we're
processing the last. If new identical events come in continuously,
faster than they're being processed, the buffer will grow forever
(though as soon as a different message comes in or there is a pause it
will be emptied).
Check instead for reference equality so we can always skip to the
latest.

Test: I can't find any reasonable way to test this
Fixes: 185828188
Change-Id: I10a78e94bcc11e9e09c242aff3ca70c609af663e

GitOrigin-RevId: 474881062f32dbb7eec081cd3a2fc21ccc631350